Extract JetpackInstallHeaderView for reuse
